move postgres helpers into `DB`
